Cleanup util methods and duplicate synchronous methods removed
