+++ packages/generator-typescript/templates/gjs/gjs.d.ts
@@ -5,8 +5,8 @@
  * If you found a bug fix it in <%= APP_NAME %> itself or create a bug report on <%= APP_SOURCE %>
  */
 
-<%_ const GObject = dep.getSync('GObject', '2.0') _%>
-<%_ const GLib = dep.getSync('GLib', '2.0') _%>
+<%_ const GObject = await dep.get('GObject', '2.0') _%>
+<%_ const GLib = await dep.get('GLib', '2.0') _%>
 <%- GObject ? GObject.importDef : '' %>
 <%- GLib ? GLib.importDef : '' %>
